@@ -1,4 +1,3 @@
-import * as _ from "lodash";
 import {Pipe, PipeTransform} from "@angular/core";
 
 @Pipe({
@@ -8,7 +7,7 @@ export class DataFilterPipe implements PipeTransform {
 
     transform(array: any[], query: string): any {
         if (query) {
-            return _.filter(array, row=>row.name.indexOf(query) > -1);
+            return array.filter(row=>row.name.indexOf(query) > -1);
         }
         return array;
     }
