@@ -3,7 +3,8 @@ import {SimpleChange, Component} from "@angular/core";
 import {DataTable, PageEvent, SortEvent} from "./DataTable";
 import {TestBed, ComponentFixture} from "@angular/core/testing";
 import {By} from "@angular/platform-browser";
-import * as _ from "lodash";
+import {switchMap} from 'rxjs/operators';
+import {range} from "rxjs";
 
 @Component({
     template: `<table [mfData]="[]"></table>`
@@ -29,13 +30,13 @@ describe("DataTable directive tests", ()=> {
             {id: 5, name: 'Ðrone'},
             {id: 4, name: 'Ananas'}
         ];
-        datatable.ngOnChanges({inputData: new SimpleChange(null, datatable.inputData)});
+        datatable.ngOnChanges({inputData: new SimpleChange(null, datatable.inputData, false)});
     });
 
     describe("initializing", ()=> {
 
         it("data should be empty array if inputData is undefined or null", () => {
-            datatable.ngOnChanges({inputData: new SimpleChange(null, null)});
+            datatable.ngOnChanges({inputData: new SimpleChange(null, null, false)});
             datatable.ngDoCheck();
             expect(datatable.data).toEqual([]);
         });
@@ -108,12 +109,19 @@ describe("DataTable directive tests", ()=> {
             });
 
             datatable.rowsOnPage = 3;
-            datatable.ngOnChanges({rowsOnPage: new SimpleChange(2, 3)});
+            datatable.ngOnChanges({rowsOnPage: new SimpleChange(2, 3, false)});
             datatable.ngDoCheck();
             expect(datatable.data).toEqual([{id: 3, name: 'banana'}, {id: 1, name: 'Duck'}, {id: 2, name: 'ącki'}]);
-
-
         });
+
+        it("should emit a dataLength of 0 when inputData is null or undefined", (done) => {
+            datatable.onPageChange.subscribe((pageOptions: PageEvent)=> {
+                expect(pageOptions.dataLength).toEqual(0);
+                done();
+            });
+            datatable.inputData = null;
+            datatable.setPage(2, 3);
+        })
     });
 
     describe("sorting", ()=> {
@@ -128,8 +136,8 @@ describe("DataTable directive tests", ()=> {
             datatable.sortBy = "id";
             datatable.sortOrder = "asc";
             datatable.ngOnChanges({
-                sortBy: new SimpleChange(null, datatable.sortBy),
-                sortOrder: new SimpleChange(null, datatable.sortOrder)
+                sortBy: new SimpleChange(null, datatable.sortBy, false),
+                sortOrder: new SimpleChange(null, datatable.sortOrder, false)
             });
             datatable.ngDoCheck();
             expect(datatable.data).toEqual([
@@ -151,8 +159,8 @@ describe("DataTable directive tests", ()=> {
             datatable.sortBy = "id";
             datatable.sortOrder = "desc";
             datatable.ngOnChanges({
-                sortBy: new SimpleChange(null, datatable.sortBy),
-                sortOrder: new SimpleChange(null, datatable.sortOrder)
+                sortBy: new SimpleChange(null, datatable.sortBy, false),
+                sortOrder: new SimpleChange(null, datatable.sortOrder, false)
             });
             datatable.ngDoCheck();
 
@@ -167,7 +175,7 @@ describe("DataTable directive tests", ()=> {
             datatable.ngDoCheck();
             datatable.sortBy = "id";
             datatable.ngOnChanges({
-                sortBy: new SimpleChange(null, datatable.sortBy)
+                sortBy: new SimpleChange(null, datatable.sortBy, false)
             });
             datatable.ngDoCheck();
             expect(datatable.sortOrder).toEqual("asc");
@@ -181,11 +189,25 @@ describe("DataTable directive tests", ()=> {
             });
             datatable.ngDoCheck();
             datatable.sortBy = "id";
-            datatable.sortOrder = "bulb";
+            datatable.sortOrder = "bulb" as any;
             datatable.ngOnChanges({
-                sortBy: new SimpleChange(null, datatable.sortBy),
-                sortOrder: new SimpleChange(null, datatable.sortOrder)
+                sortBy: new SimpleChange(null, datatable.sortBy, false),
+                sortOrder: new SimpleChange(null, datatable.sortOrder, false)
             });
+            datatable.ngDoCheck();
+            expect(datatable.sortOrder).toEqual("asc");
+            expect(datatable.data).toEqual([
+                {id: 1, name: 'Duck'},
+                {id: 2, name: 'ącki'},
+                {id: 3, name: 'banana'},
+                {id: 4, name: 'Ananas'},
+                {id: 5, name: 'Ðrone'}
+            ]);
+        });
+
+        it("should set sortOrder to 'asc' if setSort is given something else than 'asc' or 'desc'", () => {
+            datatable.setSort("id", "bulb" as any);
+            expect(datatable.getSort()).toEqual({sortBy: "id", sortOrder: "asc"});
             datatable.ngDoCheck();
             expect(datatable.sortOrder).toEqual("asc");
             expect(datatable.data).toEqual([
@@ -204,17 +226,19 @@ describe("DataTable directive tests", ()=> {
             });
             datatable.ngDoCheck();
             datatable.sortOrder = "desc";
-            datatable.ngOnChanges({sortOrder: new SimpleChange(null, datatable.sortOrder)});
+            datatable.ngOnChanges({sortOrder: new SimpleChange(null, datatable.sortOrder, false)});
             datatable.ngDoCheck();
             expect(datatable.data).toEqual(datatable.inputData);
         });
 
         it("should call output event when sorting changed", (done)=> {
             datatable.ngDoCheck();
-            datatable.sortByChange.switchMap((sortBy: string)=> {
-                expect(sortBy).toEqual("id");
-                return datatable.sortOrderChange;
-            }).subscribe((sortOrder: string)=> {
+            datatable.sortByChange.pipe(
+                switchMap((sortBy: string)=> {
+                    expect(sortBy).toEqual("id");
+                    return datatable.sortOrderChange;
+                })
+            ).subscribe((sortOrder: string)=> {
                 expect(sortOrder).toEqual("desc");
                 done();
             });
@@ -228,8 +252,8 @@ describe("DataTable directive tests", ()=> {
                 done.fail("Shouldn't call sortOrderChange");
             });
             done();
-            datatable.sortOrder = "bulb";
-            datatable.ngOnChanges({sortOrder: new SimpleChange(null, datatable.sortOrder)});
+            datatable.sortOrder = "bulb" as any;
+            datatable.ngOnChanges({sortOrder: new SimpleChange(null, datatable.sortOrder, false)});
             datatable.ngDoCheck();
         });
         // Wywołanie outputa gdy zmiana z innej strony
@@ -272,14 +296,16 @@ describe("DataTable directive tests", ()=> {
                 {name: 'Claire', age: 9},
                 {name: 'Anna', age: 34},
                 {name: 'Claire', age: 16},
+                {name: 'Anna', age: 12},
                 {name: 'Claire', age: 7},
                 {name: 'Anna', age: 12}
             ];
-            datatable.ngOnChanges({inputData: new SimpleChange(datatable.inputData, newData)});
+            datatable.ngOnChanges({inputData: new SimpleChange(datatable.inputData, newData, false)});
             datatable.setSort(['name', 'age'], "asc");
             datatable.ngDoCheck();
 
             expect(datatable.data).toEqual([
+                {name: 'Anna', age: 12},
                 {name: 'Anna', age: 12},
                 {name: 'Anna', age: 34},
                 {name: 'Claire', age: 7},
@@ -297,7 +323,7 @@ describe("DataTable directive tests", ()=> {
                 {name: 'Claire', city: {zip: '11111'}},
                 {name: 'Anna', city: {zip: '21111'}}
             ];
-            datatable.ngOnChanges({inputData: new SimpleChange(datatable.inputData, newData)});
+            datatable.ngOnChanges({inputData: new SimpleChange(datatable.inputData, newData, false)});
             datatable.setSort("city.zip", "asc");
             datatable.ngDoCheck();
 
@@ -315,7 +341,7 @@ describe("DataTable directive tests", ()=> {
     describe("data change", ()=> {
         it("should refresh data when inputData change", ()=> {
             let newData = [{id: 5, name: 'Ðrone'}, {id: 4, name: 'Ananas'}];
-            datatable.ngOnChanges({inputData: new SimpleChange(datatable.inputData, newData)});
+            datatable.ngOnChanges({inputData: new SimpleChange(datatable.inputData, newData, false)});
             datatable.ngDoCheck();
             expect(datatable.data).toEqual([{id: 5, name: 'Ðrone'}, {id: 4, name: 'Ananas'}]);
         });
@@ -347,7 +373,7 @@ describe("DataTable directive tests", ()=> {
                 done();
             });
             let newData = [{id: 5, name: 'Ðrone'}, {id: 4, name: 'Ananas'}];
-            datatable.ngOnChanges({inputData: new SimpleChange(datatable.inputData, newData)});
+            datatable.ngOnChanges({inputData: new SimpleChange(datatable.inputData, newData, false)});
             datatable.ngDoCheck();
         });
 
@@ -375,7 +401,7 @@ describe("DataTable directive tests", ()=> {
                 expect(opt.rowsOnPage).toEqual(2);
                 done();
             });
-            _.times(3, ()=>datatable.inputData.pop());
+            range(0, 3).forEach(()=>datatable.inputData.pop());
             datatable.ngDoCheck();
         });
 
@@ -384,7 +410,7 @@ describe("DataTable directive tests", ()=> {
             datatable.ngDoCheck();
 
             let newData = [{id: 5, name: 'Ðrone'}, {id: 4, name: 'Ananas'}];
-            datatable.ngOnChanges({inputData: new SimpleChange(datatable.inputData, newData)});
+            datatable.ngOnChanges({inputData: new SimpleChange(datatable.inputData, newData, false)});
             datatable.ngDoCheck();
             expect(datatable.data).toEqual(newData);
         });
@@ -406,7 +432,7 @@ describe("DataTable directive tests", ()=> {
             datatable.ngDoCheck();
 
             let newData = [{id: 5, name: 'Ðrone'}, {id: 1, name: 'Duck'}, {id: 4, name: 'Ananas'}];
-            datatable.ngOnChanges({inputData: new SimpleChange(datatable.inputData, newData)});
+            datatable.ngOnChanges({inputData: new SimpleChange(datatable.inputData, newData, false)});
             datatable.ngDoCheck();
             expect(datatable.data).toEqual([{id: 1, name: 'Duck'}]);
         });
@@ -436,7 +462,7 @@ describe("DataTable directive tests", ()=> {
             datatable.ngDoCheck();
 
             let newData = [];
-            datatable.ngOnChanges({inputData: new SimpleChange(datatable.inputData, newData)});
+            datatable.ngOnChanges({inputData: new SimpleChange(datatable.inputData, newData, false)});
             datatable.ngDoCheck();
             expect(datatable.activePage).toEqual(1);
         });
@@ -445,7 +471,7 @@ describe("DataTable directive tests", ()=> {
             datatable.setPage(2, 1);
             datatable.ngDoCheck();
 
-            _.times(5, ()=>datatable.inputData.pop());
+            range(0, 5).forEach(()=>datatable.inputData.pop());
             datatable.ngDoCheck();
             expect(datatable.inputData.length).toEqual(0);
             expect(datatable.activePage).toEqual(1);
